Add old Alchemy Vision work
This commit contains intermediate code for Alchemy Vision. I started this last week but didn't finish or commit. I am committing it nor for reference, in case it is useful when implementing Alchemy Vision in the `master` branch.

Note that this branch probably should *not* just be merged into `master`. After referring to this branch to add code to the `master` branch, it should probably be forcibly deleted.